+++ tests/test_climate.py
@@ -136,9 +136,9 @@ async def test_cerfthermostat(factory: helper.Factory) -> None:
 
     await climate.set_hvac_mode(HvacMode.OFF)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU0000050:4",
+        channel_address="VCU0000050:4",
         paramset_key="VALUES",
-        value={"MANU_MODE": 12.0, "SET_TEMPERATURE": 4.5},
+        values={"MANU_MODE": 12.0, "SET_TEMPERATURE": 4.5},
     )
 
     assert climate.hvac_mode == HvacMode.OFF
@@ -267,18 +267,18 @@ async def test_ceipthermostat(factory: helper.Factory) -> None:
 
     await climate.set_hvac_mode(HvacMode.OFF)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1769958:1",
+        channel_address="VCU1769958:1",
         paramset_key="VALUES",
-        value={"CONTROL_MODE": 1, "SET_POINT_TEMPERATURE": 4.5},
+        values={"CONTROL_MODE": 1, "SET_POINT_TEMPERATURE": 4.5},
     )
     assert climate.hvac_mode == HvacMode.OFF
     assert climate.hvac_action == HvacAction.OFF
 
     await climate.set_hvac_mode(HvacMode.HEAT)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1769958:1",
+        channel_address="VCU1769958:1",
         paramset_key="VALUES",
-        value={"CONTROL_MODE": 1, "SET_POINT_TEMPERATURE": 5.0},
+        values={"CONTROL_MODE": 1, "SET_POINT_TEMPERATURE": 5.0},
     )
     await central.event(const.INTERFACE_ID, "VCU1769958:1", "SET_POINT_MODE", ModeHmIP.MANU.value)
     assert climate.hvac_mode == HvacMode.HEAT
@@ -297,9 +297,9 @@ async def test_ceipthermostat(factory: helper.Factory) -> None:
 
     await climate.set_hvac_mode(HvacMode.AUTO)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1769958:1",
+        channel_address="VCU1769958:1",
         paramset_key="VALUES",
-        value={"BOOST_MODE": False, "CONTROL_MODE": 0},
+        values={"BOOST_MODE": False, "CONTROL_MODE": 0},
     )
     await central.event(const.INTERFACE_ID, "VCU1769958:1", "SET_POINT_MODE", ModeHmIP.AUTO.value)
     await central.event(const.INTERFACE_ID, "VCU1769958:1", "BOOST_MODE", 1)
@@ -331,9 +331,9 @@ async def test_ceipthermostat(factory: helper.Factory) -> None:
     with freeze_time("2023-03-03 08:00:00"):
         await climate.enable_away_mode_by_duration(hours=100, away_temperature=17.0)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1769958:1",
+        channel_address="VCU1769958:1",
         paramset_key="VALUES",
-        value={
+        values={
             "SET_POINT_MODE": 2,
             "SET_POINT_TEMPERATURE": 17.0,
             "PARTY_TIME_START": "2023_03_03 07:50",
@@ -345,9 +345,9 @@ async def test_ceipthermostat(factory: helper.Factory) -> None:
         start=datetime(2000, 12, 1), end=datetime(2024, 12, 1), away_temperature=17.0
     )
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1769958:1",
+        channel_address="VCU1769958:1",
         paramset_key="VALUES",
-        value={
+        values={
             "SET_POINT_MODE": 2,
             "SET_POINT_TEMPERATURE": 17.0,
             "PARTY_TIME_START": "2000_12_01 00:00",
@@ -357,9 +357,9 @@ async def test_ceipthermostat(factory: helper.Factory) -> None:
 
     await climate.disable_away_mode()
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1769958:1",
+        channel_address="VCU1769958:1",
         paramset_key="VALUES",
-        value={
+        values={
             "SET_POINT_MODE": 2,
             "PARTY_TIME_START": "2000_01_01 00:00",
             "PARTY_TIME_END": "2000_01_01 00:00",