+++ tests/test_light.py
@@ -77,20 +77,22 @@ async def test_cedimmer(factory: helper.Factory) -> None:
 
     await light.turn_on(on_time=5.0, ramp_time=6.0)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1399816:4",
+        channel_address="VCU1399816:4",
         paramset_key="VALUES",
-        value={"LEVEL": 0.10980392156862745, "RAMP_TIME": 6.0, "ON_TIME": 5.0},
+        values={"LEVEL": 0.10980392156862745, "RAMP_TIME": 6.0, "ON_TIME": 5.0},
     )
     await light.turn_on(on_time=5.0)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1399816:4",
+        channel_address="VCU1399816:4",
         paramset_key="VALUES",
-        value={"ON_TIME": 5.0, "LEVEL": 0.10980392156862745},
+        values={"ON_TIME": 5.0, "LEVEL": 0.10980392156862745},
     )
 
     await light.turn_off(ramp_time=6.0)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1399816:4", paramset_key="VALUES", value={"RAMP_TIME": 6.0, "LEVEL": 0.0}
+        channel_address="VCU1399816:4",
+        paramset_key="VALUES",
+        values={"RAMP_TIME": 6.0, "LEVEL": 0.0},
     )
     assert light.brightness == 0
     await light.turn_on()
@@ -104,7 +106,9 @@ async def test_cedimmer(factory: helper.Factory) -> None:
     light.set_on_time(0.5)
     await light.turn_on()
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1399816:4", paramset_key="VALUES", value={"ON_TIME": 0.5, "LEVEL": 1.0}
+        channel_address="VCU1399816:4",
+        paramset_key="VALUES",
+        values={"ON_TIME": 0.5, "LEVEL": 1.0},
     )
 
     await light.turn_on()
@@ -307,7 +311,7 @@ async def test_ceipfixedcolorlight(factory: helper.Factory) -> None:
     assert light.channel_hs_color is None
     await light.turn_on()
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8", paramset_key="VALUES", value={"COLOR": 7, "LEVEL": 1.0}
+        channel_address="VCU3716619:8", paramset_key="VALUES", values={"COLOR": 7, "LEVEL": 1.0}
     )
     assert light.brightness == 255
     await light.turn_on(brightness=28)
@@ -327,43 +331,43 @@ async def test_ceipfixedcolorlight(factory: helper.Factory) -> None:
 
     await light.turn_on(hs_color=(350, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8", paramset_key="VALUES", value={"COLOR": 4, "LEVEL": 1.0}
+        channel_address="VCU3716619:8", paramset_key="VALUES", values={"COLOR": 4, "LEVEL": 1.0}
     )
     assert light.color_name == FixedColor.RED
 
     await light.turn_on(hs_color=(0.0, 0.0))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8", paramset_key="VALUES", value={"COLOR": 7, "LEVEL": 1.0}
+        channel_address="VCU3716619:8", paramset_key="VALUES", values={"COLOR": 7, "LEVEL": 1.0}
     )
     assert light.color_name == FixedColor.WHITE
 
     await light.turn_on(hs_color=(60.0, 50.0))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8", paramset_key="VALUES", value={"COLOR": 6, "LEVEL": 1.0}
+        channel_address="VCU3716619:8", paramset_key="VALUES", values={"COLOR": 6, "LEVEL": 1.0}
     )
     assert light.color_name == FixedColor.YELLOW
 
     await light.turn_on(hs_color=(120, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8", paramset_key="VALUES", value={"COLOR": 2, "LEVEL": 1.0}
+        channel_address="VCU3716619:8", paramset_key="VALUES", values={"COLOR": 2, "LEVEL": 1.0}
     )
     assert light.color_name == FixedColor.GREEN
 
     await light.turn_on(hs_color=(180, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8", paramset_key="VALUES", value={"COLOR": 3, "LEVEL": 1.0}
+        channel_address="VCU3716619:8", paramset_key="VALUES", values={"COLOR": 3, "LEVEL": 1.0}
     )
     assert light.color_name == FixedColor.TURQUOISE
 
     await light.turn_on(hs_color=(240, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8", paramset_key="VALUES", value={"COLOR": 1, "LEVEL": 1.0}
+        channel_address="VCU3716619:8", paramset_key="VALUES", values={"COLOR": 1, "LEVEL": 1.0}
     )
     assert light.color_name == FixedColor.BLUE
 
     await light.turn_on(hs_color=(300, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8", paramset_key="VALUES", value={"COLOR": 5, "LEVEL": 1.0}
+        channel_address="VCU3716619:8", paramset_key="VALUES", values={"COLOR": 5, "LEVEL": 1.0}
     )
     assert light.color_name == FixedColor.PURPLE
 
@@ -378,47 +382,47 @@ async def test_ceipfixedcolorlight(factory: helper.Factory) -> None:
     light.set_on_time(18)
     await light.turn_on()
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8",
+        channel_address="VCU3716619:8",
         paramset_key="VALUES",
-        value={"DURATION_VALUE": 18, "LEVEL": 1.0},
+        values={"DURATION_VALUE": 18, "LEVEL": 1.0},
     )
 
     await light.turn_off()
     light.set_on_time(17000)
     await light.turn_on()
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8",
+        channel_address="VCU3716619:8",
         paramset_key="VALUES",
-        value={"DURATION_UNIT": 1, "DURATION_VALUE": 283, "LEVEL": 1.0},
+        values={"DURATION_UNIT": 1, "DURATION_VALUE": 283, "LEVEL": 1.0},
     )
 
     await light.turn_off()
     light.set_on_time(1000000)
     await light.turn_on()
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8",
+        channel_address="VCU3716619:8",
         paramset_key="VALUES",
-        value={"DURATION_UNIT": 2, "DURATION_VALUE": 277, "LEVEL": 1.0},
+        values={"DURATION_UNIT": 2, "DURATION_VALUE": 277, "LEVEL": 1.0},
     )
     await light.turn_on(ramp_time=18)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8",
+        channel_address="VCU3716619:8",
         paramset_key="VALUES",
-        value={"RAMP_TIME_VALUE": 18, "LEVEL": 1.0},
+        values={"RAMP_TIME_VALUE": 18, "LEVEL": 1.0},
     )
 
     await light.turn_on(ramp_time=17000)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8",
+        channel_address="VCU3716619:8",
         paramset_key="VALUES",
-        value={"RAMP_TIME_UNIT": 1, "RAMP_TIME_VALUE": 283, "LEVEL": 1.0},
+        values={"RAMP_TIME_UNIT": 1, "RAMP_TIME_VALUE": 283, "LEVEL": 1.0},
     )
 
     await light.turn_on(ramp_time=1000000)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU3716619:8",
+        channel_address="VCU3716619:8",
         paramset_key="VALUES",
-        value={"RAMP_TIME_UNIT": 2, "RAMP_TIME_VALUE": 277, "LEVEL": 1.0},
+        values={"RAMP_TIME_UNIT": 2, "RAMP_TIME_VALUE": 277, "LEVEL": 1.0},
     )
 
     await light.turn_on()
@@ -465,18 +469,18 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
     assert light.color_name == FixedColor.BLACK
     await light.turn_on()
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "COLOR": 7, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 1, "COLOR": 7, "LEVEL": 1.0},
     )
     assert light.brightness == 255
     assert light.color_name == FixedColor.WHITE
 
     await light.turn_on(brightness=100)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "LEVEL": 0.39215686274509803},
+        values={"COLOR_BEHAVIOUR": 1, "LEVEL": 0.39215686274509803},
     )
     assert light.brightness == 100
     assert light.color_name == FixedColor.WHITE
@@ -492,9 +496,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(hs_color=(350, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "COLOR": 4, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 1, "COLOR": 4, "LEVEL": 1.0},
     )
     assert light.brightness == 255
     assert light.color_name == FixedColor.RED
@@ -502,9 +506,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(hs_color=(0.0, 0.0))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "COLOR": 7, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 1, "COLOR": 7, "LEVEL": 1.0},
     )
     assert light.brightness == 255
     assert light.color_name == FixedColor.WHITE
@@ -512,9 +516,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(hs_color=(60.0, 50.0))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "COLOR": 6, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 1, "COLOR": 6, "LEVEL": 1.0},
     )
     assert light.brightness == 255
     assert light.color_name == FixedColor.YELLOW
@@ -522,9 +526,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(hs_color=(120, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "COLOR": 2, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 1, "COLOR": 2, "LEVEL": 1.0},
     )
     assert light.brightness == 255
     assert light.color_name == FixedColor.GREEN
@@ -532,9 +536,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(hs_color=(180, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "COLOR": 3, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 1, "COLOR": 3, "LEVEL": 1.0},
     )
     assert light.brightness == 255
     assert light.color_name == FixedColor.TURQUOISE
@@ -542,9 +546,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(hs_color=(240, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "COLOR": 1, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 1, "COLOR": 1, "LEVEL": 1.0},
     )
     assert light.brightness == 255
     assert light.color_name == FixedColor.BLUE
@@ -552,9 +556,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(hs_color=(300, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "COLOR": 5, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 1, "COLOR": 5, "LEVEL": 1.0},
     )
     assert light.brightness == 255
     assert light.color_name == FixedColor.PURPLE
@@ -567,9 +571,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(brightness=100)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "LEVEL": 0.39215686274509803},
+        values={"COLOR_BEHAVIOUR": 1, "LEVEL": 0.39215686274509803},
     )
     assert light.brightness == 100
     assert light.color_name == FixedColor.PURPLE
@@ -577,9 +581,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(brightness=33)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 1, "LEVEL": 0.12941176470588237},
+        values={"COLOR_BEHAVIOUR": 1, "LEVEL": 0.12941176470588237},
     )
     assert light.brightness == 33
     assert light.color_name == FixedColor.PURPLE
@@ -587,9 +591,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(effect="FLASH_MIDDLE")
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 6, "LEVEL": 0.12941176470588237},
+        values={"COLOR_BEHAVIOUR": 6, "LEVEL": 0.12941176470588237},
     )
     assert light.brightness == 33
     assert light.color_name == FixedColor.PURPLE
@@ -597,9 +601,9 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
 
     await light.turn_on(brightness=66)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 6, "LEVEL": 0.25882352941176473},
+        values={"COLOR_BEHAVIOUR": 6, "LEVEL": 0.25882352941176473},
     )
     assert light.brightness == 66
     assert light.color_name == FixedColor.PURPLE
@@ -610,47 +614,47 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
     light.set_on_time(18)
     await light.turn_on()
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 6, "DURATION_VALUE": 18, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 6, "DURATION_VALUE": 18, "LEVEL": 1.0},
     )
 
     await light.turn_off()
     light.set_on_time(17000)
     await light.turn_on()
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 6, "DURATION_UNIT": 1, "DURATION_VALUE": 283, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 6, "DURATION_UNIT": 1, "DURATION_VALUE": 283, "LEVEL": 1.0},
     )
 
     await light.turn_off()
     light.set_on_time(1000000)
     await light.turn_on()
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 6, "DURATION_UNIT": 2, "DURATION_VALUE": 277, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 6, "DURATION_UNIT": 2, "DURATION_VALUE": 277, "LEVEL": 1.0},
     )
     await light.turn_on(ramp_time=18)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 6, "RAMP_TIME_VALUE": 18, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 6, "RAMP_TIME_VALUE": 18, "LEVEL": 1.0},
     )
 
     await light.turn_on(ramp_time=17000)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 6, "RAMP_TIME_UNIT": 1, "RAMP_TIME_VALUE": 283, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 6, "RAMP_TIME_UNIT": 1, "RAMP_TIME_VALUE": 283, "LEVEL": 1.0},
     )
 
     await light.turn_on(ramp_time=1000000)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 6, "RAMP_TIME_UNIT": 2, "RAMP_TIME_VALUE": 277, "LEVEL": 1.0},
+        values={"COLOR_BEHAVIOUR": 6, "RAMP_TIME_UNIT": 2, "RAMP_TIME_VALUE": 277, "LEVEL": 1.0},
     )
 
     await light.turn_on()
@@ -666,15 +670,17 @@ async def test_ceipfixedcolorlightwired(factory: helper.Factory) -> None:
     await light.turn_off()
     await light.turn_on(effect="BLINKING_SLOW")
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8", paramset_key="VALUES", value={"COLOR_BEHAVIOUR": 2, "LEVEL": 1.0}
+        channel_address="VCU4704397:8",
+        paramset_key="VALUES",
+        values={"COLOR_BEHAVIOUR": 2, "LEVEL": 1.0},
     )
 
     await light.turn_on(brightness=28)
     await light.turn_on(effect="FLASH_MIDDLE")
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU4704397:8",
+        channel_address="VCU4704397:8",
         paramset_key="VALUES",
-        value={"COLOR_BEHAVIOUR": 6, "LEVEL": 0.10980392156862745},
+        values={"COLOR_BEHAVIOUR": 6, "LEVEL": 0.10980392156862745},
     )
 
 
@@ -741,9 +747,9 @@ async def test_ceiprgbwlight(factory: helper.Factory) -> None:
     assert light.color_temp is None
     await light.turn_on(color_temp=300)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU5629873:1",
+        channel_address="VCU5629873:1",
         paramset_key="VALUES",
-        value={"COLOR_TEMPERATURE": 3333, "LEVEL": 1.0},
+        values={"COLOR_TEMPERATURE": 3333, "LEVEL": 1.0},
     )
     assert light.color_temp == 300
 
@@ -760,30 +766,30 @@ async def test_ceiprgbwlight(factory: helper.Factory) -> None:
     assert light.hs_color is None
     await light.turn_on(hs_color=(44.4, 69.3))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU5629873:1",
+        channel_address="VCU5629873:1",
         paramset_key="VALUES",
-        value={"HUE": 44, "SATURATION": 0.693, "LEVEL": 1.0},
+        values={"HUE": 44, "SATURATION": 0.693, "LEVEL": 1.0},
     )
     assert light.hs_color == (44, 69.3)
 
     await light.turn_on(hs_color=(0, 50))
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU5629873:1",
+        channel_address="VCU5629873:1",
         paramset_key="VALUES",
-        value={"HUE": 0, "SATURATION": 0.5, "LEVEL": 1.0},
+        values={"HUE": 0, "SATURATION": 0.5, "LEVEL": 1.0},
     )
     assert light.hs_color == (0.0, 50.0)
 
     await light.turn_on(effect="EFFECT_01_END_CURRENT_PROFILE")
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU5629873:1", paramset_key="VALUES", value={"EFFECT": 1, "LEVEL": 1.0}
+        channel_address="VCU5629873:1", paramset_key="VALUES", values={"EFFECT": 1, "LEVEL": 1.0}
     )
 
     await light.turn_on(hs_color=(44, 66), ramp_time=5)
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU5629873:1",
+        channel_address="VCU5629873:1",
         paramset_key=ParamsetKey.VALUES,
-        value={
+        values={
             "HUE": 44,
             "SATURATION": 0.66,
             "DURATION_VALUE": 111600,