+++ changelog.md
@@ -3,6 +3,7 @@
 - Make entity event async
 - Extract looper from central and reuse for json/xml_rpc
 - Move loop_check to async_support
+- Add last value send cache
 
 # Version 2024.4.7 (2024-04-13)